+++ engine/internal/aggregate/controller.go
@@ -10,7 +10,7 @@ import (
 	"github.com/dogmatiq/testkit/engine/internal/panicx"
 	"github.com/dogmatiq/testkit/envelope"
 	"github.com/dogmatiq/testkit/fact"
-	"github.com/dogmatiq/testkit/internal/location"
+	"github.com/dogmatiq/testkit/location"
 )
 
 // Controller is an implementation of engine.Controller for