+++ CHANGELOG.md
@@ -18,6 +18,7 @@ should continue to work without modification.
 ### Added
 
 - **[BC]** Added `PredicateOptions` to `PredicateScope`
+- **[BC]** Added `Action.Location()` method
 
 ### Changed